Linq and database
Ok, so I've created a linq class for the news page, which I've renamed
'homenews', because with linq, it cuts of the last letter of a table,
and 'new' is a keyword, so that's no good. There's an error (line 40 and
60) in the linqClass page because 'date' is a timestamp, which I've read
is the best data type to use for when information is posted (instead of
datetime). Anyways, that's pretty much the extent of what I'm going to
do until we meet up next week and determine the best way to go forth.
The only other thing I'll try to do is create a guest username and
password we can all use so we can all access the database on my godaddy
account.